+++ packages/stargate/src/signingstargateclient.spec.ts
@@ -56,7 +56,7 @@ describe("SigningStargateClient", () => {
       const registry = new Registry();
       registry.register("/custom.MsgCustom", MsgSend);
       const options: SigningStargateClientOptions = { ...defaultSigningClientOptions, registry: registry };
-      const client = await SigningStargateClient.connectWithSigner(simapp.tendermintUrl, wallet, options);
+      const client = await SigningStargateClient.connectWithSigner(simapp.tendermintUrlHttp, wallet, options);
       const openedClient = client as unknown as PrivateSigningStargateClient;
       expect(openedClient.registry.lookupType("/custom.MsgCustom")).toEqual(MsgSend);
     });
@@ -67,7 +67,7 @@ describe("SigningStargateClient", () => {
       pendingWithoutSimapp();
       const wallet = await DirectSecp256k1HdWallet.fromMnemonic(faucet.mnemonic);
       const client = await SigningStargateClient.connectWithSigner(
-        simapp.tendermintUrl,
+        simapp.tendermintUrlHttp,
         wallet,
         defaultSigningClientOptions,
       );
@@ -95,7 +95,7 @@ describe("SigningStargateClient", () => {
       pendingWithoutSimapp();
       const wallet = await DirectSecp256k1HdWallet.fromMnemonic(faucet.mnemonic);
       const client = await SigningStargateClient.connectWithSigner(
-        simapp.tendermintUrl,
+        simapp.tendermintUrlHttp,
         wallet,
         defaultSigningClientOptions,
       );
@@ -136,7 +136,7 @@ describe("SigningStargateClient", () => {
       pendingWithoutSimapp();
       const wallet = await Secp256k1HdWallet.fromMnemonic(faucet.mnemonic);
       const client = await SigningStargateClient.connectWithSigner(
-        simapp.tendermintUrl,
+        simapp.tendermintUrlHttp,
         wallet,
         defaultSigningClientOptions,
       );
@@ -180,12 +180,12 @@ describe("SigningStargateClient", () => {
       });
       const [{ address: signer }, { address: payer }] = await wallet.getAccounts();
       const client = await SigningStargateClient.connectWithSigner(
-        simapp.tendermintUrl,
+        simapp.tendermintUrlHttp,
         wallet,
         defaultSigningClientOptions,
       );
 
-      const cometClient = await Tendermint34Client.connect(simapp.tendermintUrl);
+      const cometClient = await Tendermint34Client.connect(simapp.tendermintUrlHttp);
       const queryClient = QueryClient.withExtensions(cometClient, setupFeegrantExtension);
       let allowanceExists: boolean;
       try {
@@ -254,7 +254,7 @@ describe("SigningStargateClient", () => {
       pending("We cannot test this easily anymore since the IBC module was removed from simapp");
       const wallet = await DirectSecp256k1HdWallet.fromMnemonic(faucet.mnemonic);
       const client = await SigningStargateClient.connectWithSigner(
-        simapp.tendermintUrl,
+        simapp.tendermintUrlHttp,
         wallet,
         defaultSigningClientOptions,
       );
@@ -303,7 +303,7 @@ describe("SigningStargateClient", () => {
       pending("We cannot test this easily anymore since the IBC module was removed from simapp");
       const wallet = await Secp256k1HdWallet.fromMnemonic(faucet.mnemonic);
       const client = await SigningStargateClient.connectWithSigner(
-        simapp.tendermintUrl,
+        simapp.tendermintUrlHttp,
         wallet,
         defaultSigningClientOptions,
       );
@@ -355,7 +355,7 @@ describe("SigningStargateClient", () => {
         pendingWithoutSimapp();
         const wallet = await DirectSecp256k1HdWallet.fromMnemonic(faucet.mnemonic);
         const client = await SigningStargateClient.connectWithSigner(
-          simapp.tendermintUrl,
+          simapp.tendermintUrlHttp,
           wallet,
           defaultSigningClientOptions,
         );
@@ -386,7 +386,7 @@ describe("SigningStargateClient", () => {
         pendingWithoutSimapp();
         const wallet = await DirectSecp256k1HdWallet.fromMnemonic(faucet.mnemonic);
         const client = await SigningStargateClient.connectWithSigner(
-          simapp.tendermintUrl,
+          simapp.tendermintUrlHttp,
           wallet,
           defaultSigningClientOptions,
         );
@@ -415,7 +415,7 @@ describe("SigningStargateClient", () => {
       it("works with auto gas", async () => {
         pendingWithoutSimapp();
         const wallet = await DirectSecp256k1HdWallet.fromMnemonic(faucet.mnemonic);
-        const client = await SigningStargateClient.connectWithSigner(simapp.tendermintUrl, wallet, {
+        const client = await SigningStargateClient.connectWithSigner(simapp.tendermintUrlHttp, wallet, {
           ...defaultSigningClientOptions,
           gasPrice: defaultGasPrice,
         });
@@ -437,7 +437,7 @@ describe("SigningStargateClient", () => {
         pendingWithoutSimapp();
         const wallet = await ModifyingDirectSecp256k1HdWallet.fromMnemonic(faucet.mnemonic);
         const client = await SigningStargateClient.connectWithSigner(
-          simapp.tendermintUrl,
+          simapp.tendermintUrlHttp,
           wallet,
           defaultSigningClientOptions,
         );
@@ -476,7 +476,7 @@ describe("SigningStargateClient", () => {
         pendingWithoutSimapp();
         const wallet = await Secp256k1HdWallet.fromMnemonic(faucet.mnemonic);
         const client = await SigningStargateClient.connectWithSigner(
-          simapp.tendermintUrl,
+          simapp.tendermintUrlHttp,
           wallet,
           defaultSigningClientOptions,
         );
@@ -503,7 +503,7 @@ describe("SigningStargateClient", () => {
         pendingWithoutSimapp();
         const wallet = await Secp256k1HdWallet.fromMnemonic(faucet.mnemonic);
         const client = await SigningStargateClient.connectWithSigner(
-          simapp.tendermintUrl,
+          simapp.tendermintUrlHttp,
           wallet,
           defaultSigningClientOptions,
         );
@@ -530,7 +530,7 @@ describe("SigningStargateClient", () => {
         pendingWithoutSimapp();
         const wallet = await Secp256k1HdWallet.fromMnemonic(faucet.mnemonic);
         const client = await SigningStargateClient.connectWithSigner(
-          simapp.tendermintUrl,
+          simapp.tendermintUrlHttp,
           wallet,
           defaultSigningClientOptions,
         );
@@ -641,7 +641,11 @@ describe("SigningStargateClient", () => {
           registry: customRegistry,
           aminoTypes: customAminoTypes,
         };
-        const client = await SigningStargateClient.connectWithSigner(simapp.tendermintUrl, wallet, options);
+        const client = await SigningStargateClient.connectWithSigner(
+          simapp.tendermintUrlHttp,
+          wallet,
+          options,
+        );
 
         const msg: CustomMsgDelegate = {
           customDelegatorAddress: faucet.address0,
@@ -665,7 +669,7 @@ describe("SigningStargateClient", () => {
         pendingWithoutSimapp();
         const wallet = await ModifyingSecp256k1HdWallet.fromMnemonic(faucet.mnemonic);
         const client = await SigningStargateClient.connectWithSigner(
-          simapp.tendermintUrl,
+          simapp.tendermintUrlHttp,
           wallet,
           defaultSigningClientOptions,
         );
@@ -706,7 +710,7 @@ describe("SigningStargateClient", () => {
         pendingWithoutSimapp();
         const wallet = await DirectSecp256k1HdWallet.fromMnemonic(faucet.mnemonic);
         const client = await SigningStargateClient.connectWithSigner(
-          simapp.tendermintUrl,
+          simapp.tendermintUrlHttp,
           wallet,
           defaultSigningClientOptions,
         );
@@ -736,7 +740,7 @@ describe("SigningStargateClient", () => {
       it("works with auto gas", async () => {
         pendingWithoutSimapp();
         const wallet = await DirectSecp256k1HdWallet.fromMnemonic(faucet.mnemonic);
-        const client = await SigningStargateClient.connectWithSigner(simapp.tendermintUrl, wallet, {
+        const client = await SigningStargateClient.connectWithSigner(simapp.tendermintUrlHttp, wallet, {
           ...defaultSigningClientOptions,
           gasPrice: defaultGasPrice,
         });
@@ -762,7 +766,7 @@ describe("SigningStargateClient", () => {
         pendingWithoutSimapp();
         const wallet = await ModifyingDirectSecp256k1HdWallet.fromMnemonic(faucet.mnemonic);
         const client = await SigningStargateClient.connectWithSigner(
-          simapp.tendermintUrl,
+          simapp.tendermintUrlHttp,
           wallet,
           defaultSigningClientOptions,
         );
@@ -796,7 +800,7 @@ describe("SigningStargateClient", () => {
         pendingWithoutSimapp();
         const wallet = await Secp256k1HdWallet.fromMnemonic(faucet.mnemonic);
         const client = await SigningStargateClient.connectWithSigner(
-          simapp.tendermintUrl,
+          simapp.tendermintUrlHttp,
           wallet,
           defaultSigningClientOptions,
         );
@@ -826,7 +830,7 @@ describe("SigningStargateClient", () => {
         pendingWithoutSimapp();
         const wallet = await Secp256k1HdWallet.fromMnemonic(faucet.mnemonic);
         const client = await SigningStargateClient.connectWithSigner(
-          simapp.tendermintUrl,
+          simapp.tendermintUrlHttp,
           wallet,
           defaultSigningClientOptions,
         );
@@ -860,7 +864,7 @@ describe("SigningStargateClient", () => {
         pendingWithoutSimapp();
         const wallet = await DirectSecp256k1HdWallet.fromMnemonic(faucet.mnemonic);
         const client = await SigningStargateClient.connectWithSigner(
-          simapp.tendermintUrl,
+          simapp.tendermintUrlHttp,
           wallet,
           defaultSigningClientOptions,
         );
@@ -890,7 +894,7 @@ describe("SigningStargateClient", () => {
         pendingWithoutSimapp();
         const wallet = await ModifyingDirectSecp256k1HdWallet.fromMnemonic(faucet.mnemonic);
         const client = await SigningStargateClient.connectWithSigner(
-          simapp.tendermintUrl,
+          simapp.tendermintUrlHttp,
           wallet,
           defaultSigningClientOptions,
         );
@@ -927,7 +931,7 @@ describe("SigningStargateClient", () => {
         pendingWithoutSimapp();
         const wallet = await DirectSecp256k1HdWallet.fromMnemonic(faucet.mnemonic);
         const client = await SigningStargateClient.connectWithSigner(
-          simapp.tendermintUrl,
+          simapp.tendermintUrlHttp,
           wallet,
           defaultSigningClientOptions,
         );
@@ -958,7 +962,7 @@ describe("SigningStargateClient", () => {
         pendingWithoutSimapp();
         const wallet = await DirectSecp256k1HdWallet.fromMnemonic(faucet.mnemonic);
         const client = await SigningStargateClient.connectWithSigner(
-          simapp.tendermintUrl,
+          simapp.tendermintUrlHttp,
           wallet,
           defaultSigningClientOptions,
         );
@@ -997,7 +1001,7 @@ describe("SigningStargateClient", () => {
         pendingWithoutSimapp();
         const wallet = await Secp256k1HdWallet.fromMnemonic(faucet.mnemonic);
         const client = await SigningStargateClient.connectWithSigner(
-          simapp.tendermintUrl,
+          simapp.tendermintUrlHttp,
           wallet,
           defaultSigningClientOptions,
         );
@@ -1027,7 +1031,7 @@ describe("SigningStargateClient", () => {
         pendingWithoutSimapp();
         const wallet = await Secp256k1HdWallet.fromMnemonic(faucet.mnemonic);
         const client = await SigningStargateClient.connectWithSigner(
-          simapp.tendermintUrl,
+          simapp.tendermintUrlHttp,
           wallet,
           defaultSigningClientOptions,
         );
@@ -1141,7 +1145,11 @@ describe("SigningStargateClient", () => {
           registry: customRegistry,
           aminoTypes: customAminoTypes,
         };
-        const client = await SigningStargateClient.connectWithSigner(simapp.tendermintUrl, wallet, options);
+        const client = await SigningStargateClient.connectWithSigner(
+          simapp.tendermintUrlHttp,
+          wallet,
+          options,
+        );
 
         const msg: CustomMsgDelegate = {
           customDelegatorAddress: faucet.address0,
@@ -1168,7 +1176,7 @@ describe("SigningStargateClient", () => {
         pendingWithoutSimapp();
         const wallet = await ModifyingSecp256k1HdWallet.fromMnemonic(faucet.mnemonic);
         const client = await SigningStargateClient.connectWithSigner(
-          simapp.tendermintUrl,
+          simapp.tendermintUrlHttp,
           wallet,
           defaultSigningClientOptions,
         );
@@ -1205,7 +1213,7 @@ describe("SigningStargateClient", () => {
         pendingWithoutSimapp();
         const wallet = await Secp256k1HdWallet.fromMnemonic(faucet.mnemonic);
         const client = await SigningStargateClient.connectWithSigner(
-          simapp.tendermintUrl,
+          simapp.tendermintUrlHttp,
           wallet,
           defaultSigningClientOptions,
         );
@@ -1236,7 +1244,7 @@ describe("SigningStargateClient", () => {
         pendingWithoutSimapp();
         const wallet = await Secp256k1HdWallet.fromMnemonic(faucet.mnemonic);
         const client = await SigningStargateClient.connectWithSigner(
-          simapp.tendermintUrl,
+          simapp.tendermintUrlHttp,
           wallet,
           defaultSigningClientOptions,
         );